HW13_Dash/requirements.txt
tensorflow
gdown
pillow
numpy
requests
matplotlib
scikit-learn
scipy
pandas
seaborn
scikit-image
opencv-python
tqdm
plotly
pydub
moviepy